converted to double precision - fourth build - succesful simulation
git-svn-id: https://svn.oss.deltares.nl/repos/imod/trunk@1866 8a8853d3-af7d-47ec-b359-4bd2096ba547

+++ MODFLOW/MF2005/gwf2met1.f
@@ -688,7 +688,7 @@ c arguments
       character(len=16), intent(in) :: text
       integer, intent(in) :: ibdchn
       integer, intent(in) :: ncol, nrow, nlay
-      real,dimension(ncol,nrow,nlay) :: buff
+      real(kind=8),dimension(ncol,nrow,nlay) :: buff
       integer, intent(in) :: iout
       logical, intent(out) :: retflag
 
@@ -790,10 +790,10 @@ c modules
 
 c arguments
       character(len=16), intent(in) :: text
-      integer, intent(in)           :: ichn
-      integer, intent(in)           :: ncol, nrow
-      real, dimension(ncol,nrow)    :: buff
-      logical, intent(out)          :: retflag
+      integer, intent(in) :: ichn
+      integer, intent(in) :: ncol, nrow
+      real(kind=8), dimension(ncol,nrow) :: buff
+      logical, intent(out) :: retflag
 
 c local variables
       character(len=1024) :: fname
